requirements.txt
beautifulsoup4 
requests
supabase
python-dotenv
playwright